All: Fix warnings reported by -Wformat
Since the tfp_printf() functions now have format attributes, the
format of the string has to match the arguments.

+++ src/main/fc/cli.c
@@ -877,7 +877,7 @@ static void cliSerialPassthrough(char *cmdline)
             tfp_printf("Port %d could not be opened.\r\n", id);
             return;
         }
-        tfp_printf("Port %d opened, baud = %d.\r\n", id, baud);
+        tfp_printf("Port %d opened, baud = %u.\r\n", id, (unsigned)baud);
     } else {
         passThroughPort = passThroughPortUsage->serialPort;
         // If the user supplied a mode, override the port's mode, otherwise
